requirements.txt
gym==0.17.2
Box2D==2.3.10 
tensorflow-gpu==1.14
mujoco-py<2.1,>=2.0
stable-baselines[mpi]==2.10.0 